requirements_tf_cpu.txt
tensorflow-cpu
matplotlib
Pillow
pydot
PyYAML
